+++ test/beast/http/read.cpp
@@ -21,7 +21,8 @@
 #include <boost/beast/test/stream.hpp>
 #include <boost/beast/test/yield_to.hpp>
 #include <boost/beast/unit_test/suite.hpp>
-#include <boost/asio/spawn.hpp>
+#include <boost/asio/io_service.hpp>
+#include <boost/asio/strand.hpp>
 #include <atomic>
 
 namespace boost {
@@ -480,6 +481,47 @@ public:
             });
     }
 
+    struct copyable_handler
+    {
+        template<class... Args>
+        void
+        operator()(Args&&...) const
+        {
+        }
+    };
+
+    void
+    testAsioHandlerInvoke()
+    {
+        // make sure things compile, also can set a
+        // breakpoint in asio_handler_invoke to make sure
+        // it is instantiated.
+        {
+            boost::asio::io_context ioc;
+            boost::asio::io_service::strand s{ioc};
+            test::stream ts{ioc};
+            flat_buffer b;
+            request_parser<dynamic_body> p;
+            async_read_some(ts, b, p, s.wrap(copyable_handler{}));
+        }
+        {
+            boost::asio::io_context ioc;
+            boost::asio::io_service::strand s{ioc};
+            test::stream ts{ioc};
+            flat_buffer b;
+            request_parser<dynamic_body> p;
+            async_read(ts, b, p, s.wrap(copyable_handler{}));
+        }
+        {
+            boost::asio::io_context ioc;
+            boost::asio::io_service::strand s{ioc};
+            test::stream ts{ioc};
+            flat_buffer b;
+            request<dynamic_body> m;
+            async_read(ts, b, m, s.wrap(copyable_handler{}));
+        }
+    }
+
     void
     run() override
     {
@@ -502,6 +544,7 @@ public:
         testIoService();
         testRegression430();
         testReadGrind();
+        testAsioHandlerInvoke();
     }
 };