+++ test/beast/websocket/read2.cpp
@@ -12,6 +12,8 @@
 
 #include "test.hpp"
 
+#include <boost/asio/io_service.hpp>
+#include <boost/asio/strand.hpp>
 #include <boost/asio/write.hpp>
 
 namespace boost {
@@ -649,6 +651,30 @@ public:
             move_only_handler{});
     }
 
+    struct copyable_handler
+    {
+        template<class... Args>
+        void
+        operator()(Args&&...) const
+        {
+        }
+    };
+
+    void
+    testAsioHandlerInvoke()
+    {
+        // make sure things compile, also can set a
+        // breakpoint in asio_handler_invoke to make sure
+        // it is instantiated.
+        {
+            boost::asio::io_context ioc;
+            boost::asio::io_service::strand s{ioc};
+            stream<test::stream> ws{ioc};
+            flat_buffer b;
+            ws.async_read(b, s.wrap(copyable_handler{}));
+        }
+    }
+
     void
     run() override
     {
@@ -660,6 +686,7 @@ public:
         testIssueBF1();
         testIssueBF2();
         testMoveOnly();
+        testAsioHandlerInvoke();
     }
 };